refactor signup views into signup file
