+++ posthog/api/signup.py
@@ -1,13 +1,17 @@
-from typing import Any, Dict, Optional, cast
+from typing import Any, Dict, Optional, Union, cast
 
 import posthoganalytics
+from django import forms
 from django.conf import settings
 from django.contrib.auth import login, password_validation
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError, transaction
-from django.shortcuts import get_object_or_404
+from django.shortcuts import redirect, render
 from django.urls.base import reverse
 from rest_framework import exceptions, generics, permissions, response, serializers, validators
+from sentry_sdk import capture_exception
+from social_core.pipeline.partial import partial
+from social_django.strategy import DjangoStrategy
 
 from posthog.api.shared import UserBasicSerializer
 from posthog.demo import create_demo_team
@@ -16,7 +20,7 @@ from posthog.models import Organization, Team, User
 from posthog.models.organization import OrganizationInvite
 from posthog.permissions import CanCreateOrg
 from posthog.tasks import user_identify
-from posthog.utils import mask_email_address
+from posthog.utils import get_can_create_org, mask_email_address
 
 
 class SignupSerializer(serializers.Serializer):
@@ -83,41 +87,10 @@ class SignupSerializer(serializers.Serializer):
         return posthoganalytics.feature_enabled("new-onboarding-2822", user.distinct_id)
 
 
-class SocialSignupSerializer(serializers.Serializer):
-    """
-    Signup serializer when the account is created using social authentication.
-    Pre-processes information not obtained from SSO provider to create organization.
-    """
-
-    organization_name: serializers.Field = serializers.CharField(max_length=128)
-    email_opt_in: serializers.Field = serializers.BooleanField(default=True)
-
-    def create(self, validated_data, **kwargs):
-        request = self.context["request"]
-
-        if not request.session.get("backend"):
-            raise serializers.ValidationError(
-                "Inactive social login session. Go to /login and log in before continuing.",
-            )
-
-        request.session["organization_name"] = validated_data["organization_name"]
-        request.session["email_opt_in"] = validated_data["email_opt_in"]
-        request.session.set_expiry(3600)  # 1 hour to complete process
-        return {"continue_url": reverse("social:complete", args=[request.session["backend"]])}
-
-    def to_representation(self, instance: Any) -> Any:
-        return self.instance
-
-
 class SignupViewset(generics.CreateAPIView):
     serializer_class = SignupSerializer
     # Enables E2E testing of signup flow
     permission_classes = (permissions.AllowAny,) if settings.E2E_TESTING else (CanCreateOrg,)
-
-
-class SocialSignupViewset(generics.CreateAPIView):
-    serializer_class = SocialSignupSerializer
-    permission_classes = (CanCreateOrg,)
 
 
 class InviteSignupSerializer(serializers.Serializer):
@@ -234,3 +207,143 @@ class InviteSignupViewset(generics.CreateAPIView):
                 "organization_name": invite.organization.name,
             }
         )
+
+
+## Social Signup
+## views & serializers
+class SocialSignupSerializer(serializers.Serializer):
+    """
+    Signup serializer when the account is created using social authentication.
+    Pre-processes information not obtained from SSO provider to create organization.
+    """
+
+    organization_name: serializers.Field = serializers.CharField(max_length=128)
+    email_opt_in: serializers.Field = serializers.BooleanField(default=True)
+
+    def create(self, validated_data, **kwargs):
+        request = self.context["request"]
+
+        if not request.session.get("backend"):
+            raise serializers.ValidationError(
+                "Inactive social login session. Go to /login and log in before continuing.",
+            )
+
+        request.session["organization_name"] = validated_data["organization_name"]
+        request.session["email_opt_in"] = validated_data["email_opt_in"]
+        request.session.set_expiry(3600)  # 1 hour to complete process
+        return {"continue_url": reverse("social:complete", args=[request.session["backend"]])}
+
+    def to_representation(self, instance: Any) -> Any:
+        return self.instance
+
+
+class SocialSignupViewset(generics.CreateAPIView):
+    serializer_class = SocialSignupSerializer
+    permission_classes = (CanCreateOrg,)
+
+
+class TeamInviteSurrogate:
+    """This reimplements parts of OrganizationInvite that enable compatibility with the old Team.signup_token."""
+
+    def __init__(self, signup_token: str):
+        team = Team.objects.select_related("organization").get(signup_token=signup_token)
+        self.organization = team.organization
+
+    def validate(*args, **kwargs) -> bool:
+        return True
+
+    def use(self, user: Any, *args, **kwargs) -> None:
+        user.join(organization=self.organization)
+
+
+class CompanyNameForm(forms.Form):
+    companyName = forms.CharField(max_length=64)
+    emailOptIn = forms.BooleanField(required=False)
+
+
+def finish_social_signup(request):
+    """
+    TODO: DEPRECATED in favor of posthog.api.signup.SocialSignupSerializer
+    """
+    if not get_can_create_org():
+        return redirect("/login?error=no_new_organizations")
+
+    if request.method == "POST":
+        form = CompanyNameForm(request.POST)
+        if form.is_valid():
+            request.session["organization_name"] = form.cleaned_data["companyName"]
+            request.session["email_opt_in"] = bool(form.cleaned_data["emailOptIn"])
+            return redirect(reverse("social:complete", args=[request.session["backend"]]))
+    else:
+        form = CompanyNameForm()
+    return render(request, "signup_to_organization_company.html", {"user_name": request.session["user_name"]})
+
+
+@partial
+def social_create_user(strategy: DjangoStrategy, details, backend, request, user=None, *args, **kwargs):
+    if user:
+        return {"is_new": False}
+    user_email = details["email"][0] if isinstance(details["email"], (list, tuple)) else details["email"]
+    user_name = details["fullname"] or details["username"]
+    strategy.session_set("user_name", user_name)
+    strategy.session_set("backend", backend.name)
+    from_invite = False
+    invite_id = strategy.session_get("invite_id")
+    if not invite_id:
+        organization_name = strategy.session_get("organization_name", None)
+        email_opt_in = strategy.session_get("email_opt_in", None)
+        if not organization_name or email_opt_in is None:
+            return redirect(finish_social_signup)
+
+        serializer = SignupSerializer(
+            data={
+                "organization_name": organization_name,
+                "email_opt_in": email_opt_in,
+                "first_name": user_name,
+                "email": user_email,
+                "password": None,
+            },
+            context={"request": request},
+        )
+
+        serializer.is_valid(raise_exception=True)
+        user = serializer.save()
+    else:
+        from_invite = True
+        try:
+            invite: Union[OrganizationInvite, TeamInviteSurrogate] = OrganizationInvite.objects.select_related(
+                "organization",
+            ).get(id=invite_id)
+        except (OrganizationInvite.DoesNotExist, ValidationError):
+            try:
+                invite = TeamInviteSurrogate(invite_id)
+            except Team.DoesNotExist:
+                return redirect(f"/signup/{invite_id}?error_code=invalid_invite&source=social_create_user")
+
+        try:
+            invite.validate(user=None, email=user_email)
+        except exceptions.ValidationError as e:
+            return redirect(
+                f"/signup/{invite_id}?error_code={e.get_codes()[0]}&error_detail={e.args[0]}&source=social_create_user"
+            )
+
+        try:
+            user = strategy.create_user(email=user_email, first_name=user_name, password=None)
+        except Exception as e:
+            capture_exception(e)
+            message = "Account unable to be created. This account may already exist. Please try again"
+            " or use different credentials."
+            return redirect(f"/signup/{invite_id}?error_code=unknown&error_detail={message}&source=social_create_user")
+
+        invite.use(user, prevalidated=True)
+
+    report_user_signed_up(
+        distinct_id=user.distinct_id,
+        is_instance_first_user=User.objects.count() == 1,
+        is_organization_first_user=not from_invite,
+        new_onboarding_enabled=False,
+        backend_processor="social_create_user",
+        social_provider=backend.name,
+    )
+
+    return {"is_new": True, "user": user}
